Convert id type from int to long

+++ src/com/excilys/formation/cdb/dao/CompanyDaoImpl.java
@@ -14,7 +14,7 @@ import com.excilys.formation.cdb.persistence.ConnectionManager;
 public class CompanyDaoImpl implements CompanyDao {
 
 	@Override
-	public List<Company> list(int id_first, int nb_to_print) {
+	public List<Company> list(long id_first, int nb_to_print) {
 		Connection conn = ConnectionManager.INSTANCE.getConnection();
 		PreparedStatement ps = null;
 		ResultSet rs = null;
@@ -22,7 +22,7 @@ public class CompanyDaoImpl implements CompanyDao {
 		
 		try {
 			ps = conn.prepareStatement("SELECT * FROM company WHERE id >= ? LIMIT ?");
-			ps.setInt(1, id_first);
+			ps.setLong(1, id_first);
 			ps.setInt(2, nb_to_print);
 			rs = ps.executeQuery();
 			
@@ -40,7 +40,7 @@ public class CompanyDaoImpl implements CompanyDao {
 	}
 
 	@Override
-	public List<Company> list(int id_first) {
+	public List<Company> list(long id_first) {
 		return this.list(id_first, 10);
 	}
